+++ NEWS.md
@@ -8,13 +8,18 @@
 
 - Added new boostrap functionality for models with constant or variable returns to scale based on Simar and Wilson (1998) (#6)
 - Added new function `unset_env_vars()` that will clear environment variables set by pioneeR if the app exits unexpectedly and the variables are not unset automatically
+- Added new function `compute_scale_efficiency()` to run CRS, VRS and NIRS models to compute scale efficiency and evaluate optimal scale size for each unit
+- Added new function `summary_tbl_dea()` to create a summary table of efficiency scores from a vector of efficiency scores or an object of type Farrell as returned from the `dea()` function in the Benchmarking package
+- Added new helper function `create_matrix()` that will create a matrix of input values or output values based on a data.frame
 
 ## Improvements
 
-- Layout changed to `layout_sidebar` from the bslib package to improve UI and fix compatability with bslib >= 0.6.0 and Boostrap 5.3 (#44)
+- Layout changed to `layout_sidebar()` from the bslib package to improve UI and fix compatability with bslib >= 0.6.0 and Boostrap 5.3 (#44)
+- Analysis report has been updated and simplified to make the results from the app easier to reproduce (#35)
 - `runPioneeR()` has been renamed `run_pioneer()`. `runPioneeR()` will still work, but will be deprecated in a later version
+- Simplified logic in PDF report generated from the analysis tab
 
-# Bug fixes
+## Bug fixes
 
 - If the user selects a port number that is considered unsafe, a random port number is selected instead (#40)
 - Fixed a bug where an error occurred if any of the model parameters were not found in the column names (#39)